requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
asgiref==3.10.0
attrs==25.4.0
bleach==6.2.0
boto==2.49.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
defusedxml==0.7.1
dj-database-url==3.0.1
Django==5.2.7
django-blog-it==0.3.1
django-ckeditor==6.7.1
django-ckeditor-5==0.2.13
django-cors-headers==4.9.0
django-filter==25.2
django-js-asset==3.1.2
django-simple-pagination==1.3
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
frozenlist==1.8.0
geoip2==4.8.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
httplib2==0.31.0
idna==3.11
inflection==0.5.1
maxminddb==2.8.2
microurl==0.1.1
multidict==6.7.0
numpy==2.2.5
oauthlib==3.3.1
packaging==25.0
pandas==2.3.3
pillow==12.0.0
propcache==0.4.1
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-decouple==3.8
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
social-auth-app-django==5.6.0
social-auth-core==4.8.1
sqlparse==0.5.3
tinycss2==1.4.0
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
webencodings==0.5.1
yarl==1.22.0
